Universe and lattice power calculation now working properly in Python, but multithreading is now broken

+++ src/Cell.cpp
@@ -1,7 +1,7 @@
 #include "Cell.h"
 
 
-short int Cell::_n = 0;
+int Cell::_n = 0;
 
 static int auto_id = 10000;
 
@@ -34,7 +34,7 @@ Cell::Cell() {
  * @param id the user-specified cell ID
  * @param universe the ID of the universe within which this cell resides
  */
-Cell::Cell(short int universe, short int id) {
+Cell::Cell(int universe, int id) {
 
     /* If the user did not define an optional ID, create one */
     if (id == 0)
@@ -65,7 +65,7 @@ Cell::~Cell() {
  * @brief Return the cell's unique ID.
  * @return the cell's unique ID
  */
-short int Cell::getUid() const {
+int Cell::getUid() const {
     return _uid;
 }
 
@@ -74,7 +74,7 @@ short int Cell::getUid() const {
  * @brief Return the cell's user-specified ID.
  * @return the cell's user-specified ID
  */
-short int Cell::getId() const {
+int Cell::getId() const {
     return _id;
 }
 
@@ -92,7 +92,7 @@ cellType Cell::getType() const {
  * @brief Return the ID of the universe within which this cell resides.
  * @return the universe ID
  */
-short int Cell::getUniverse() const {
+int Cell::getUniverse() const {
     return _universe;
 }
 
@@ -101,7 +101,7 @@ short int Cell::getUniverse() const {
  * @brief Return the number of surfaces in the cell.
  * @return the number of surfaces
  */
-short int Cell::getNumSurfaces() const {
+int Cell::getNumSurfaces() const {
     return this->_surfaces.size();
 }
 
@@ -111,7 +111,7 @@ short int Cell::getNumSurfaces() const {
  *        surfaces making up the cell.
  * @return map of surface IDs and surface pointers
  */
-std::map<short int,Surface*> Cell::getSurfaces() const {
+std::map<int,Surface*> Cell::getSurfaces() const {
     return _surfaces;
 }
 
@@ -120,7 +120,7 @@ std::map<short int,Surface*> Cell::getSurfaces() const {
  * @brief Set the ID for teh universe within whic this cell resides.
  * @param universe the universe's user-specified ID
  */
-void Cell::setUniverse(short int universe) {
+void Cell::setUniverse(int universe) {
     _universe = universe;
 }
 
@@ -130,13 +130,13 @@ void Cell::setUniverse(short int universe) {
  * @param halfspace the surface halfspace (positive/negative for surface side)
  * @param surface a pointer to the surface
  */
-void Cell::addSurface(short int halfspace, Surface* surface) {
+void Cell::addSurface(int halfspace, Surface* surface) {
     if (halfspace != -1 && halfspace != +1)
         log_printf(ERROR, "Unable to add surface %d to cell %d since the "
                     "halfspace %d is not -1 or 1", surface->getId(), 
                                                     _id, halfspace);
 
-    _surfaces.insert(std::pair<short int, Surface*>(halfspace*surface->getId(),
+    _surfaces.insert(std::pair<int, Surface*>(halfspace*surface->getId(),
                                                                      surface));
 }
 
@@ -187,7 +187,7 @@ void Cell::setSurfacePointer(Surface* surface) {
 bool Cell::cellContainsPoint(Point* point) {
 
     /* Loop over all surfaces inside the cell */
-    std::map<short int, Surface*>::iterator iter;
+    std::map<int, Surface*>::iterator iter;
     for (iter = _surfaces.begin(); iter != _surfaces.end(); ++iter) {
       
         /* If the surface evaluated point is not the same sign as the surface
@@ -227,7 +227,7 @@ double Cell::minSurfaceDist(Point* point, double angle,
     double d;
     Point intersection;
 
-    std::map<short int, Surface*>::iterator iter;
+    std::map<int, Surface*>::iterator iter;
     
     /* Loop over all of the cell's surfaces */
     for (iter = _surfaces.begin(); iter != _surfaces.end(); ++iter) {
@@ -257,8 +257,8 @@ double Cell::minSurfaceDist(Point* point, double angle,
  *        (the default is zero)
  * @param id the user-specified cell ID
  */
-CellBasic::CellBasic(short int universe, short int material, int rings, 
-		     int sectors, short int id): Cell(universe, id) {
+CellBasic::CellBasic(int universe, int material, int rings, 
+		     int sectors, int id): Cell(universe, id) {
     _cell_type = MATERIAL;
     _material = material;
     setNumRings(rings);
@@ -271,7 +271,7 @@ CellBasic::CellBasic(short int universe, short int material, int rings,
  * @brief Return the ID of the material filling the cellbasic.
  * @return the material's ID
  */
-short int CellBasic::getMaterial() const {
+int CellBasic::getMaterial() const {
     return _material;
 }
 
@@ -280,7 +280,7 @@ short int CellBasic::getMaterial() const {
  * @brief Return the number of rings in the cell.
  * @return the number of rings
  */
-short int CellBasic::getNumRings() {
+int CellBasic::getNumRings() {
     return this->_num_rings;
 }
 
@@ -289,7 +289,7 @@ short int CellBasic::getNumRings() {
  * @brief Return the number of sectors in the cell.
  * @return the number of sectors
  */
-short int CellBasic::getNumSectors() {
+int CellBasic::getNumSectors() {
     return this->_num_sectors;
 }
 
@@ -311,7 +311,7 @@ int CellBasic::getNumFSRs() {
  * @brief Set the cell's number of rings.
  * @param num_rings the number of rings in this cell
  */
-void CellBasic::setNumRings(short int num_rings) {
+void CellBasic::setNumRings(int num_rings) {
     if (num_rings < 0)
         log_printf(ERROR, "Unable to give %d rings to cell %d since this is "
 		 "a negative number", num_rings, _id);
@@ -324,7 +324,7 @@ void CellBasic::setNumRings(short int num_rings) {
  * @brief Set the cell's number of sectors.
  * @param num_sectors the number of sectors in this cell
  */
-void CellBasic::setNumSectors(short int num_sectors) {
+void CellBasic::setNumSectors(int num_sectors) {
     if (num_sectors < 0)
         log_printf(ERROR, "Unable to give %d sectors to cell %d since this is "
 		 "a negative number", num_sectors, _id);
@@ -347,7 +347,7 @@ CellBasic* CellBasic::clone() {
 					_num_rings, _num_sectors);
 
     /* Loop over all of this cell's surfaces and add them to the clone */
-    std::map<short int, Surface*>::iterator iter;
+    std::map<int, Surface*>::iterator iter;
     int halfspace;
     Surface* surface;
 
@@ -432,15 +432,15 @@ void CellBasic::ringify() {
     double y1 = 0.;
     double x2 = 0.;
     double y2 = 0.;
-    short int halfspace1 = 0;
-    short int halfspace2 = 0;
+    int halfspace1 = 0;
+    int halfspace2 = 0;
     std::vector<Circle*> circles;
 
     /* See if the cell contains 1 or 2 surfaces */
-    std::map<short int, Surface*>::iterator iter1;
+    std::map<int, Surface*>::iterator iter1;
     for (iter1=_surfaces.begin(); iter1 != _surfaces.end(); ++iter1) {
         if ((*iter1).second->getSurfaceType() == CIRCLE) {
-	    short int halfspace = (*iter1).first / (*iter1).second->getId();
+	    int halfspace = (*iter1).first / (*iter1).second->getId();
 	    Circle* circle = static_cast<Circle*>((*iter1).second);
 
 	    /* Outermost bounding circle */
@@ -599,7 +599,7 @@ std::string CellBasic::toString() {
            << ", num of sectors = " << _num_sectors;
     
     /* Append each of the surface ids to the string */
-    std::map<short int, Surface*>::iterator iter;
+    std::map<int, Surface*>::iterator iter;
     string << ", surface ids = ";
     for (iter = _surfaces.begin(); iter != _surfaces.end(); ++iter)
         string << iter->first << ", ";
@@ -623,7 +623,7 @@ void CellBasic::printString() {
  *  @param universe the ID of the universe within which this cell resides
  *  @param universe_fill the ID of the universe filling this cell
  */
-CellFill::CellFill(short int universe, short int universe_fill, short int id):
+CellFill::CellFill(int universe, int universe_fill, int id):
   Cell(universe, id) {
         _cell_type = FILL;
         _universe_fill.first = universe_fill;
@@ -634,7 +634,7 @@ CellFill::CellFill(short int universe, short int universe_fill, short int id):
  * @brief Return the ID of the universe filling this cell.
  * @return the universe's id
  */
-short int CellFill::getUniverseFillId() const {
+int CellFill::getUniverseFillId() const {
     return _universe_fill.first;
 }
 
@@ -667,7 +667,7 @@ int CellFill::getNumFSRs() {
  * @brief Set the ID of the universe filling this cellfill.
  * @param universe_fill the universe's ID
  */
-void CellFill::setUniverseFill(short int universe_fill) {
+void CellFill::setUniverseFill(int universe_fill) {
     _universe_fill.first = universe_fill;
 }
 
@@ -694,7 +694,7 @@ std::string CellFill::toString() {
       ", num_surfaces = " << getNumSurfaces();
 
     /** Add the IDs for the surfaces in this cell */
-    std::map<short int, Surface*>::iterator iter;
+    std::map<int, Surface*>::iterator iter;
     string << ", surface ids = ";
     for (iter = _surfaces.begin(); iter != _surfaces.end(); ++iter)
         string << iter->first << ", ";
